% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoskaicius.R
\name{autoskaicius}
\alias{autoskaicius}
\title{autoskaicius
autoskaicius
This function}
\usage{
autoskaicius(n, p, mydata1, mydata2)
}
\arguments{
\item{n}{A number}

\item{p}{A number}

\item{mydata1}{A list}

\item{mydata2}{A list}
}
\description{
autoskaicius
autoskaicius
This function
}
\examples{
1
}
\keyword{autoskaicius}
