% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoskaicius1.R
\name{autoskaicius1}
\alias{autoskaicius1}
\title{This function}
\usage{
autoskaicius1(n, p, mydata1, mydata2)
}
\arguments{
\item{n}{A number}

\item{p}{A number}

\item{mydata1}{A list}

\item{mydata2}{A list}
}
\description{
This function
}
\examples{
1
}
\keyword{autoskaicius1}
