% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoskaicius2.R
\name{autoskaicius2}
\alias{autoskaicius2}
\title{This function}
\usage{
autoskaicius2(masina, klientaisort)
}
\arguments{
\item{masina}{A matrix}

\item{klientaisort}{A matrix}
}
\description{
This function
}
\examples{
1
}
\keyword{autoskaicius2}
