% Generated by roxygen2: do not edit by hand
	% Please edit documentation in R/klientulentele.R
	\name{klientulentele}
	\alias{klientulentele}
	\title{This function}
	\usage{
	klientulentele()
	}
	\arguments{
	\item{}{A list}
	}
	\description{
	This function
	}
	\examples{
	1
	}
	\keyword{klientulentele}
