% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraunam.R
\name{kraunam}
\alias{kraunam}
\title{This function}
\usage{
kraunam(p, n, veid, mydata2, mydata1)
}
\arguments{
\item{p}{A number}

\item{n}{A number}

\item{veid}{A matrix}

\item{mydata2}{A matrix}

\item{mydata1}{A matrix}
}
\description{
This function
}
\examples{
1
}
\keyword{kraunam}
