% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{link}
\alias{link}
\title{This function}
\usage{
link(marsrutas, klientai, mydata2)
}
\arguments{
\item{marsrutas}{A matrix}

\item{klientai}{A matrix}

\item{mydata2}{A matrix}
}
\description{
This function
}
\examples{
1
}
\keyword{link}
