% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps_koord.R
\name{maps_koord}
\alias{maps_koord}
\title{This function}
\usage{
maps_koord(lon, lat)
}
\arguments{
\item{lon}{A number}

\item{lat}{A number}
}
\description{
This function
}
\examples{
1
}
\keyword{maps_koord}
