% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps_start.R
\name{maps_start}
\alias{maps_start}
\title{This function}
\usage{
maps_start(p, n, mydata1, mydata2, veid)
}
\arguments{
\item{p}{A number}

\item{n}{A number}

\item{mydata1}{A list}

\item{mydata2}{A list}

\item{veid}{A matrix}
}
\description{
This function
}
\examples{
1
}
\keyword{maps_start}
