% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps_test1.R
\name{maps_test1}
\alias{maps_test1}
\title{This function}
\usage{
maps_test1(p, n, mydata1, mydata2, veid)
}
\arguments{
\item{p}{A number}

\item{n}{A number}

\item{mydata1}{A matrix}

\item{mydata2}{A matrix}

\item{veid}{A matrix}
}
\description{
This function
}
\examples{
1
}
\keyword{maps_test1}
