% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paskirstymas.R
\name{paskirstymas}
\alias{paskirstymas}
\title{This function}
\usage{
paskirstymas(p, n, mydata1, mydata2, veid)
}
\arguments{
\item{p}{A number}

\item{n}{A number}

\item{mydata1}{A list}

\item{mydata2}{A list}

\item{veid}{A matrix}
}
\description{
This function
}
\examples{
1
}
\keyword{paskirstymas}
