% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rek_auto_sk.R
\name{rek_auto_sk}
\alias{rek_auto_sk}
\title{This function}
\usage{
rek_auto_sk(n, p, mydata1, mydata2)
}
\arguments{
\item{n}{A number}

\item{p}{A number}

\item{mydata1}{A matrix}

\item{mydata2}{A matrix}
}
\description{
This function
}
\examples{
1
}
\keyword{rek_auto_sk}
