% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rek_auto_sk1.R
\name{rek_auto_sk1}
\alias{rek_auto_sk1}
\title{This function}
\usage{
rek_auto_sk1(n, p, mydata1, mydata2)
}
\arguments{
\item{n}{A Number}

\item{p}{A Number}

\item{mydata1}{A matrix}

\item{mydata2}{A matrix}
}
\description{
This function
}
\examples{
1
}
\keyword{rek_auto_sk1}
