% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route1.R
\name{route1}
\alias{route1}
\title{This function}
\usage{
route1(veid, z, n, mas2)
}
\arguments{
\item{veid}{A matrix}

\item{z}{A number}

\item{n}{A number}

\item{mas2}{A matrix}
}
\description{
This function
}
\examples{
1
}
\keyword{route1}
