% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veidr.R
\name{veidr}
\alias{veidr}
\title{This function}
\usage{
veidr(veid, z, n)
}
\arguments{
\item{veid}{A matrix}

\item{z}{A number}

\item{n}{A number}
}
\description{
This function
}
\examples{
1
}
\keyword{veidr}
